Move Interval to its own internal module
